medisupply-client-service/requirements.txt
# FastAPI and web server
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# Database
sqlalchemy>=2.0.30,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.12.1,<2.0.0

# Security - JWT token validation
python-jose[cryptography]>=3.3.0,<4.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# HTTP client for NIT validation
httpx>=0.25.0,<1.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0

